refactor: move Resolver out of refs into dependencies::resolve
`refs` was both the pure protocol/tx grammar AND `Resolver`, which
needed `crate::config`. That created a refs <-> config cycle and pulled
config knowledge into the otherwise-pure vocabulary module.

Resolution (parsed ref -> project | declared dependency) is
dependency-domain logic — it queries `config.dependencies` — so move
`Resolver`/`ResolvedProtocol`/`ResolveError` into a new
`dependencies::resolve` (re-exported from `dependencies`).

Result: `refs` is now a pure leaf module with zero `crate::` deps;
the only remaining edge is the clean one-directional `config -> refs`
(DependencyEntry.reference: ProtocolRef). Strict layering:
refs -> {config, oci} -> dependencies. `oci` stays an independent
shared push/pull transport (publish + dependencies as peers).

Consumers updated: inspect/tir.rs, invoke.rs import Resolver/
ResolvedProtocol from `dependencies` now. Behavior unchanged; also
collapsed the carried-over nested if (let-chain).
